fix(website): keep the docs index Run link honest after the resolver swap; drop dead workspace helpers
- resolveDocsWorkspace never returns null (a miss resolves docs-only), so
  the docs index's Run href guard is now gated on resolution.kind ===
  'mapped' instead of a dead null check, moved into a lib module with a
  unit test that would have caught the regression
- drop dead ?? fallbacks on the now-nonnullable WorkspaceIdentity.docsPath
  in workspace-presentation.ts
- delete the unused getWebsiteWorkspaceHref helper
- fix WebsiteWorkspace.spec.tsx tests that override docsSlot: undefined
  even though the docs route they model always supplies a slot
